services/cv-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
